+++ src/main/java/hotspot/user/common/exception/code/FamilyUsageErrorCode.java
@@ -1,8 +1,9 @@
 package hotspot.user.common.exception.code;
 
+import org.springframework.http.HttpStatus;
+
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 
 @Getter
 @RequiredArgsConstructor
