+++ src/main/java/hotspot/user/usage/familyUsage/domain/FamilySubUsage.java
@@ -1,6 +1,6 @@
 package hotspot.user.usage.familyUsage.domain;
 
-import hotspot.user.usage.familyUsage.infrastructure.util.RedisUsageCalculator;
+import hotspot.user.common.util.redis.RedisUsageCalculator;
 
 public class FamilySubUsage {
