+++ src/main/java/hotspot/user/usage/familyUsage/infrastructure/respository/FamilyUsageRepositoryImpl.java
@@ -1,12 +1,10 @@
-package hotspot.user.usage.familyUsage.infrastructure;
+package hotspot.user.usage.familyUsage.infrastructure.respository;
 
 import java.util.List;
 
 import org.springframework.stereotype.Repository;
 
-import hotspot.user.usage.familyUsage.controller.response.FamilyUsageResponse;
 import hotspot.user.usage.familyUsage.domain.FamilyUsage;
-import hotspot.user.usage.familyUsage.domain.mapper.FamilyUsageMapper;
 import hotspot.user.usage.familyUsage.service.port.FamilyUsageRepository;
 import hotspot.user.usage.familyUsage.service.schema.FamilySubList;
 import lombok.RequiredArgsConstructor;
